Correct integration test synchronisation
